requirements.txt
streamlit>=1.36
google-genai>=0.3.0
Pillow>=10.3.0
